+++ lib/modules/perf/perf_service.dart
@@ -1,9 +1,9 @@
 ﻿// Module: perf
-// Updated at: 2026-08-28 17:06:57
-// Commit: perf(perf): optimize memory consumption in perf list viewport
+// Updated at: 2026-08-29 09:35:38
+// Commit: fix(perf): prevent double tap duplicate action in perf
 
 class perfComponentState {
-  final String commitRef = '3993';
-  final String lastMessage = 'perf(perf): optimize memory consumption in perf list viewport';
-  final DateTime timestamp = DateTime.parse('2026-08-28T17:06:57+05:00');
+  final String commitRef = '4006';
+  final String lastMessage = 'fix(perf): prevent double tap duplicate action in perf';
+  final DateTime timestamp = DateTime.parse('2026-08-29T09:35:38+05:00');
 }